Adding WorkerPool testing shared stuff
